+++ modules/cli_commands/get.js
@@ -1,3 +1,5 @@
 module.exports = {
-   command: 'get'
+   command: 'get',
+   alias: ['find'],
+   callback: () => {}
 }