+++ modules/cli_commands/schematic-tof.js
@@ -1,0 +1,6 @@
+module.exports = {
+   commands: '', // String
+   alias: [], // Array
+   mode: '', // String
+   callback: () => {} // Function
+}