+++ modules/commandline_system.mjs
@@ -2,9 +2,10 @@ import fs from 'fs'
 import { createRequire } from'module'
 const require = createRequire(import.meta.url)
 
-const commandCache = {}
+let commandReference = {}
+let commandCache = []
 
-export const fileLoader = () => {
+export const fileLoader = (fileToReload = null, assignID = null) => {
    const scan = dir => {
       for(const item of fs.readdirSync(dir)){
          if(fs.lstatSync(`${dir}/${item}`).isDirectory()){
@@ -12,7 +13,8 @@ export const fileLoader = () => {
             continue
          }
 
-         if(!item.endsWith('.js'))continue
+         // "-tof.js" = Test-Only-Files.js
+         if(!item.endsWith('.js') || item.endsWith('-tof.js'))continue
 
          try{
             const jsFile = require(`../${dir}/${item}`)
@@ -24,7 +26,7 @@ export const fileLoader = () => {
                delete require.cache[require.resolve(`../${dir}/${item}`)]
             }
 
-            const { command, disabled } = jsFile
+            const { command, disabled, alias } = jsFile
 
             if(!command || typeof command !== 'string' || command === ''){
                console.log('\x1b[1m\x1b[31m%s\x1b[0m', `[WARNING]: "${item}" rejected due to command header not meeting the requirements.`)
@@ -35,22 +37,31 @@ export const fileLoader = () => {
             if(disabled){
                console.log('\x1b[1m\x1b[34m%s\x1b[0m', `[INFO]: ${command} Command disabled.`)
                removeRequireCache()
-               commandCache[command] = {
-                  disabled: true
-               }
                continue
             }
 
+            commandReference[command] = commandCache.length
+
+            if(jsFile.alias)for(const alias of jsFile.alias)commandReference[alias] = commandCache.length
+
+            const commandNames = [ command ]
+
             const commandData = {
-               commandPath: `../${dir}/${item}`
+               commandNames,
+               commandFilePath: `../${dir}/${item}`
             }
+
+            for(const alt of alias)commandNames.push(alt)
 
             for(const entry of Object.entries(jsFile)){
-               if(entry[0] === 'command' || entry[0] === 'disabled')continue
+               if(entry[0] === 'command' || entry[0] === 'disabled' || entry[0] === 'alias')continue
 
-               commandData[entry[0]] = entry[1]
+               commandData[entry[0]] += entry[1]
             }
-            commandCache[command] = commandData
+
+            console.log(jsFile, '\n', commandData)
+            console.table(commandReference)
+            commandCache.push(commandData)
 
          }catch(error){
             console.log(`There was an error when loading file "${item}".`)
@@ -58,9 +69,54 @@ export const fileLoader = () => {
          }
       }
    }
+
    scan('./modules/cli_commands')
+
+   console.log(commandCache)
 }
 
-export const commandHandler = () => {}
+export const commandHandler = request => {
+   if(typeof request !== 'string')return console.log('Command is not of type string')
 
-export const reloadCommands = () => {}
+   const requestArray = request.split(/[ ]/g)
+   const commandName = requestArray[0]
+   requestArray.shift()
+   const args = requestArray
+
+   if(commandName === '')return console.log('Command is empty')
+
+   console.log("\nCommand execute request: " + request + '\x1b[1m\x1b[34m%s\x1b[0m', '\n[INFO]: Finding command...')
+
+   console.log(commandCache[commandReference[commandName]])
+}
+
+export const reloadCommands = resetCommand => {
+   const resetRequire = module => {
+      console.log('\x1b[1m\x1b[33m%s\x1b[0m', '[ALERT]: Deleting required module: ' + module)
+      delete require.cache[require.resolve(module)]
+   }
+
+   if(resetCommand === 'all'){
+      for(const command of commandCache)resetRequire(command.commandFilePath)
+
+      console.log('\x1b[1m\x1b[33m%s\x1b[0m', '[ALERT]: Resetting cache...')
+
+      commandCache = []
+      commandReference = []
+
+      console.log('\x1b[1m\x1b[34m%s\x1b[0m', '[INFO]: Reloading commands...')
+
+      fileLoader()
+      return
+   }
+
+   if(!commandReference[resetCommand])return console.log('Command not found.')
+
+   const currentVersion = commandCache[commandReference[resetCommand]]
+   const currentReferenceID = commandReference[resetCommand]
+
+   delete commandCache[commandReference[resetCommand]]
+   for(const commandName of currentVersion.commandNames)delete commandReference[commandName]
+
+   // fileLoader(currentVersion.commandFilePath, currentReferenceID)
+}